Move `TokenBundleSizeAssessment` to `Cardano.CoinSelection.Size`.
The `TokenBundleSizeAssessment` type exists solely to give the coin
selection algorithm a simple way to determine if a `TokenBundle` is too
large (which would indicate that the bundle needs to be split up).

Callers of coin selection must guarantee to provide an implementation of
`TokenBundleSizeAssessor` that satisfies the laws specified in this
type's documentation.

As such, this type should really be defined and exported as part of the
public API of `cardano-coin-selection`.
